refactor: re-org class
