+++ src/core/InstructionService.js
@@ -1,11 +1,5 @@
 import { clusterApiUrl, Connection } from "@solana/web3.js"
-import {
-  PublicKey,
-  TransactionInstruction,
-  SystemProgram,
-  SYSVAR_RENT_PUBKEY,
-} from "@solana/web3.js"
-import { Token, TOKEN_PROGRAM_ID, ASSOCIATED_TOKEN_PROGRAM_ID } from "@solana/spl-token"
+import { PublicKey, TransactionInstruction, SystemProgram } from "@solana/web3.js"
 import { hexToBytes } from "web3-utils"
 import ab2str from "arraybuffer-to-string"
 import { NEON_TOKEN_MINT, NEON_EVM_LOADER_ID } from "../constants"
@@ -59,15 +53,6 @@ class InstructionService {
     return neonAccount
   }
 
-  async getAuthorityPoolAddress() {
-    const enc = new TextEncoder()
-    const authority = await PublicKey.findProgramAddress(
-      [enc.encode("Deposit")],
-      new PublicKey(NEON_EVM_LOADER_ID),
-    )
-    return authority
-  }
-
   _getSolanaWalletPubkey() {
     return new PublicKey(this.solanaWalletAddress)
   }
@@ -90,58 +75,6 @@ class InstructionService {
 
   _getNeonMintTokenPubkey() {
     return this._getSolanaPubkey(NEON_TOKEN_MINT)
-  }
-
-  async _getERC20WrapperAddress(splToken) {
-    const enc = new TextEncoder()
-    const tokenPubkey = this._getSolanaPubkey(splToken.address_spl)
-    const erc20Seed = this._getEthSeed(splToken.address)
-    const accountSeed = this._getNeonAccountSeed()
-    const erc20addr = await PublicKey.findProgramAddress(
-      [
-        new Uint8Array([1]),
-        enc.encode("ERC20Balance"),
-        tokenPubkey.toBytes(),
-        erc20Seed,
-        accountSeed,
-      ],
-      new PublicKey(NEON_EVM_LOADER_ID),
-    )
-    return { erc20Address: erc20addr[0], erc20Nonce: erc20addr[1] }
-  }
-
-  async _getERC20WrapperAccount(splToken) {
-    const { erc20Address } = await this._getERC20WrapperAddress(splToken)
-    const ERC20WrapperAccount = await this.connection.getAccountInfo(erc20Address)
-    return ERC20WrapperAccount
-  }
-
-  async _createERC20AccountInstruction(splToken) {
-    const data = new Uint8Array([0x0f])
-    const solanaPubkey = this._getSolanaWalletPubkey()
-    const mintPublicKey = this._getSolanaPubkey(splToken.address_spl)
-    const { erc20Address } = await this._getERC20WrapperAddress(splToken)
-    const { neonAddress } = await this._getNeonAccountAddress()
-    const contractAddress = await PublicKey.findProgramAddress(
-      [new Uint8Array([1]), this._getEthSeed(splToken.address)],
-      new PublicKey(NEON_EVM_LOADER_ID),
-    )
-    const keys = [
-      { pubkey: solanaPubkey, isSigner: true, isWritable: true },
-      { pubkey: erc20Address, isSigner: false, isWritable: true },
-      { pubkey: neonAddress, isSigner: false, isWritable: true },
-      { pubkey: contractAddress[0], isSigner: false, isWritable: true },
-      { pubkey: mintPublicKey, isSigner: false, isWritable: true },
-      { pubkey: SystemProgram.programId, isSigner: false, isWritable: false },
-      { pubkey: TOKEN_PROGRAM_ID, isSigner: false, isWritable: false },
-      { pubkey: SYSVAR_RENT_PUBKEY, isSigner: false, isWritable: false },
-    ]
-    const instruction = new TransactionInstruction({
-      programId: new PublicKey(NEON_EVM_LOADER_ID),
-      data,
-      keys,
-    })
-    return instruction
   }
 
   async _getNeonAccountInstructionKeys(neonAddress = "") {
@@ -187,26 +120,6 @@ class InstructionService {
       value: "0x00", // Only required to send ether to the recipient from the initiating external account.
       data: this._computeWithdrawEthTransactionData(amount, token),
     }
-  }
-
-  async _createTransferInstruction(amount, splToken, toSolana = false) {
-    const mintPubkey = this._getSolanaPubkey(splToken.address_spl)
-    const solanaPubkey = this._getSolanaWalletPubkey()
-    const { erc20Address } = await this._getERC20WrapperAddress(splToken)
-    const solanaBalanceAccount = await Token.getAssociatedTokenAddress(
-      ASSOCIATED_TOKEN_PROGRAM_ID,
-      TOKEN_PROGRAM_ID,
-      mintPubkey,
-      solanaPubkey,
-    )
-    return Token.createTransferInstruction(
-      TOKEN_PROGRAM_ID,
-      toSolana ? erc20Address : solanaBalanceAccount,
-      toSolana ? solanaBalanceAccount : erc20Address,
-      solanaPubkey,
-      [],
-      Number(amount) * Math.pow(10, splToken.decimals),
-    )
   }
 }
