+++ src/core/NeonPortal.js
@@ -4,70 +4,9 @@ import { Transaction, TransactionInstruction, PublicKey } from "@solana/web3.js"
 import ab2str from "arraybuffer-to-string"
 import { NEON_EVM_LOADER_ID, NEON_TOKEN_DECIMALS, NEON_TOKEN_MINT } from "../constants"
 
+// Neon-token
 class NeonPortal extends InstructionService {
-  async _createApproveDepositInstruction(amount) {
-    const solanaPubkey = this._getSolanaPubkey()
-    const source = await Token.getAssociatedTokenAddress(
-      ASSOCIATED_TOKEN_PROGRAM_ID,
-      TOKEN_PROGRAM_ID,
-      new PublicKey(NEON_TOKEN_MINT),
-      solanaPubkey,
-    )
-    const [authority] = await this.getAuthorityPoolAddress()
-
-    const instruction = Token.createApproveInstruction(
-      TOKEN_PROGRAM_ID,
-      source,
-      authority,
-      solanaPubkey,
-      [],
-      Number(amount) * Math.pow(10, NEON_TOKEN_DECIMALS),
-    )
-
-    return instruction
-  }
-
-  async _createDepositTransferInstruction(amount, splToken) {
-    const mintPubkey = this._getNeonMintTokenPubkey()
-    const solanaPubkey = this._getSolanaWalletPubkey()
-    const { erc20Address } = await this._getERC20WrapperAddress(splToken)
-    const solanaBalanceAccount = await Token.getAssociatedTokenAddress(
-      ASSOCIATED_TOKEN_PROGRAM_ID,
-      TOKEN_PROGRAM_ID,
-      mintPubkey,
-      solanaPubkey,
-    )
-    return Token.createTransferInstruction(
-      TOKEN_PROGRAM_ID,
-      solanaBalanceAccount,
-      erc20Address,
-      solanaPubkey,
-      [],
-      Number(amount) * Math.pow(10, splToken.decimals),
-    )
-  }
-
-  async _createWithdrawTransferInstruction(amount, splToken) {
-    const mintPubkey = this._getSolanaPubkey(splToken.address_spl)
-    const solanaPubkey = this._getSolanaWalletPubkey()
-    const { erc20Address } = await this._getERC20WrapperAddress(splToken)
-    const solanaBalanceAccount = await Token.getAssociatedTokenAddress(
-      ASSOCIATED_TOKEN_PROGRAM_ID,
-      TOKEN_PROGRAM_ID,
-      mintPubkey,
-      solanaPubkey,
-    )
-    return Token.createTransferInstruction(
-      TOKEN_PROGRAM_ID,
-      erc20Address,
-      solanaBalanceAccount,
-      solanaPubkey,
-      [],
-      Number(amount) * Math.pow(10, splToken.decimals),
-    )
-  }
-
-  // #region
+  // #region Neon -> Solana
   async createNeonTransfer(events = undefined, amount = 0) {
     events = events === undefined ? this.events : events
     if (typeof events.onBeforeCreateInstruction === "function") events.onBeforeCreateInstruction()
@@ -97,7 +36,7 @@ class NeonPortal extends InstructionService {
       new PublicKey(NEON_TOKEN_MINT),
       solanaPubkey,
     )
-    const [authority] = await this.getAuthorityPoolAddress()
+    const [authority] = await this._getAuthorityPoolAddress()
     const pool = await Token.getAssociatedTokenAddress(
       ASSOCIATED_TOKEN_PROGRAM_ID,
       TOKEN_PROGRAM_ID,
@@ -135,27 +74,40 @@ class NeonPortal extends InstructionService {
     })
   }
 
-  _computeWithdrawEthTransactionData() {
-    const withdrawMethodID = "0x8e19899e"
+  async _createApproveDepositInstruction(amount) {
     const solanaPubkey = this._getSolanaPubkey()
-    const solanaStr = ab2str(solanaPubkey.toBytes(), "hex")
-    return `${withdrawMethodID}${solanaStr}`
+    const source = await Token.getAssociatedTokenAddress(
+      ASSOCIATED_TOKEN_PROGRAM_ID,
+      TOKEN_PROGRAM_ID,
+      new PublicKey(NEON_TOKEN_MINT),
+      solanaPubkey,
+    )
+    const [authority] = await this._getAuthorityPoolAddress()
+
+    const instruction = Token.createApproveInstruction(
+      TOKEN_PROGRAM_ID,
+      source,
+      authority,
+      solanaPubkey,
+      [],
+      Number(amount) * Math.pow(10, NEON_TOKEN_DECIMALS),
+    )
+
+    return instruction
   }
 
-  _computeWithdrawAmountValue(amount, splToken) {
-    const result = Number(amount) * Math.pow(10, splToken.decimals)
-    return "0x" + result.toString(16)
+  async _getAuthorityPoolAddress() {
+    const enc = new TextEncoder()
+    const authority = await PublicKey.findProgramAddress(
+      [enc.encode("Deposit")],
+      new PublicKey(NEON_EVM_LOADER_ID),
+    )
+    return authority
   }
 
-  getEthereumTransactionParams(amount, token) {
-    return {
-      to: "0x053e3d1b12726f648B2e45CEAbDF9078B742576D",
-      from: this.neonWalletAddress,
-      value: this._computeWithdrawAmountValue(amount, token),
-      data: this._computeWithdrawEthTransactionData(),
-    }
-  }
+  // #endregion
 
+  // #region Solana -> Neon
   async createSolanaTransfer(
     events = undefined,
     amount = 0,
@@ -181,6 +133,28 @@ class NeonPortal extends InstructionService {
       if (typeof events.onErrorSign === "function") events.onErrorSign(e)
     }
   }
+
+  getEthereumTransactionParams(amount, token) {
+    return {
+      to: "0x053e3d1b12726f648B2e45CEAbDF9078B742576D",
+      from: this.neonWalletAddress,
+      value: this._computeWithdrawAmountValue(amount, token),
+      data: this._computeWithdrawEthTransactionData(),
+    }
+  }
+
+  _computeWithdrawEthTransactionData() {
+    const withdrawMethodID = "0x8e19899e"
+    const solanaPubkey = this._getSolanaPubkey()
+    const solanaStr = ab2str(solanaPubkey.toBytes(), "hex")
+    return `${withdrawMethodID}${solanaStr}`
+  }
+
+  _computeWithdrawAmountValue(amount, splToken) {
+    const result = Number(amount) * Math.pow(10, splToken.decimals)
+    return "0x" + result.toString(16)
+  }
+  // #endregion
 }
 
 export default NeonPortal
